riverdrivertest: move benchmarks to own file
I meant to do this as part of #904 because the `riverdrivertest.go` file
is already super long, and the last thing we need is to add another
growing group of benchmarks to it.
